+++ test/integration/suite_test.go
@@ -11,10 +11,6 @@ import (
 // Testing Vars & Consts
 // -----------------------------------------------------------------------------
 
-const (
-	enterpriseLicenseEnvVar = "KONG_ENTERPRISE_LICENSE"
-)
-
 var (
 	// ctx is a common context that can be used between tests
 	ctx = context.Background()